+++ epam-promo/components/buttons/IconButton.tsx
@@ -1,26 +1,22 @@
 import { createSkinComponent } from '@epam/uui-core';
-import { allEpamPrimaryColors, EpamPrimaryColor } from '../types';
-import { IconButtonCoreProps, IconButton as uuiIconButton } from '@epam/uui';
+import { EpamPrimaryColor } from '../types';
+import * as uui from '@epam/uui';
 
-export type IconColor = EpamPrimaryColor | 'gray30' | 'gray50' | 'gray60';
-export const allIconColors: IconColor[] = [
-    ...allEpamPrimaryColors, 'gray30', 'gray50', 'gray60',
-];
+/** Defines component color. */
+type IconColor = EpamPrimaryColor | 'gray30' | 'gray50' | 'gray60';
 
-export interface IconButtonMods {
+interface IconButtonMods {
     /**
+     * Defines component color.
      * @default 'gray60'
      */
     color?: IconColor;
 }
 
-export type IconButtonProps = IconButtonCoreProps & IconButtonMods;
+/** Represents the properties for the IconButton component. */
+export type IconButtonProps = uui.IconButtonCoreProps & IconButtonMods;
 
-export const IconButton = createSkinComponent<IconButtonCoreProps, IconButtonProps>(
-    uuiIconButton,
-    (props): IconButtonProps =>
-        ({
-            color: props.color ?? 'gray60',
-        }),
-    () => [],
+export const IconButton = createSkinComponent<uui.IconButtonProps, IconButtonProps>(
+    uui.IconButton,
+    (props) => ({ color: props.color ?? 'gray60' }),
 );