+++ epam-promo/components/widgets/Tag.tsx
@@ -1,14 +1,19 @@
 import { createSkinComponent } from '@epam/uui-core';
-import { Tag as UuiTag, TagCoreProps as UuiTagCoreProps } from '@epam/uui';
+import * as uui from '@epam/uui';
 import { EpamPrimaryColor } from '../types';
 
-export type TagMods = {
+type TagMods = {
     /**
+     * Defines component color.
      * @default 'gray30'
      */
     color?: EpamPrimaryColor | 'white' | 'gray10' | 'gray30' | 'gray70';
 };
 
-export type TagProps = UuiTagCoreProps & TagMods;
+/** Represents the properties for the Tag component. */
+export type TagProps = uui.TagCoreProps & TagMods;
 
-export const Tag = createSkinComponent<UuiTagCoreProps, TagMods>(UuiTag);
+export const Tag = createSkinComponent<uui.TagProps, TagProps>(
+    uui.Tag,
+    (props) => ({ color: props.color || 'gray30' }),
+);