+++ uui-core/src/data/processing/views/BaseListView.tsx
@@ -14,6 +14,7 @@ import {
     ScrollToConfig,
 } from '../../../types';
 import { ITree, NOT_FOUND_RECORD } from './tree/ITree';
+import { flushSync } from 'react-dom';
 
 interface NodeStats {
     isSomeCheckable: boolean;
@@ -43,11 +44,17 @@ export abstract class BaseListView<TItem, TId, TFilter> implements IDataSourceVi
     abstract getVisibleRows(): DataRowProps<TItem, TId>[];
     abstract getListProps(): DataSourceListProps;
     _forceUpdate = () => {
-        !this.isDestroyed && this.onValueChange({ ...this.value });
+        if (!this.isDestroyed) {
+            flushSync(() => { this.onValueChange({ ...this.value }); });
+        }
     };
 
-    public destroy() {
+    public deactivate() {
         this.isDestroyed = true;
+    }
+
+    public activate() {
+        this.isDestroyed = false;
     }
 
     protected constructor(editable: IEditable<DataSourceState<TFilter, TId>>, protected props: BaseListViewProps<TItem, TId, TFilter>) {
@@ -366,7 +373,8 @@ export abstract class BaseListView<TItem, TId, TFilter> implements IDataSourceVi
                             row.isChildrenChecked = row.isChildrenChecked || childStats.isSomeChecked;
                             row.isChildrenSelected = childStats.isSomeSelected;
                             stats = this.mergeStats(stats, childStats);
-                        } else if (!row.isFolded && appendRows) {
+                            // while searching and no children in visible tree, no need to append placeholders.
+                        } else if (!this.value.search && !row.isFolded && appendRows) {
                             // children are not loaded
                             const parentsWithRow = [...row.path, this.visibleTree.getPathItem(item)];
                             for (let m = 0; m < estimatedChildrenCount && rows.length < lastIndex; m++) {
@@ -610,6 +618,10 @@ export abstract class BaseListView<TItem, TId, TFilter> implements IDataSourceVi
         return count;
     };
 
+    public clearAllChecked = () => {
+        this.handleSelectAll(false);
+    };
+
     private mergeStats = (parentStats: NodeStats, childStats: NodeStats) => ({
         ...parentStats,
         isSomeCheckable: parentStats.isSomeCheckable || childStats.isSomeCheckable,
@@ -662,4 +674,13 @@ export abstract class BaseListView<TItem, TId, TFilter> implements IDataSourceVi
     protected isPartialLoad = () => false;
     public loadData() {}
     public abstract reload(): void;
+
+    public getConfig() {
+        return {
+            complexIds: this.props.complexIds,
+            cascadeSelection: this.props.cascadeSelection,
+            selectAll: this.props.selectAll,
+            backgroundReload: this.props.backgroundReload,
+        };
+    }
 }