+++ uui-core/src/types/dataSources.ts
@@ -197,16 +197,35 @@ export interface BaseListViewProps<TItem, TId, TFilter> {
     backgroundReload?: boolean;
 }
 
+export type IDataSourceViewConfig = {
+    complexIds?: boolean;
+    cascadeSelection?: CascadeSelection;
+    selectAll?: true | false;
+    backgroundReload?: boolean;
+    flattenSearchResults?: boolean;
+};
+
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 export type IDataSourceView<TItem, TId, TFilter> = {
+    getConfig(): IDataSourceViewConfig;
     getById(id: TId, index: number): DataRowProps<TItem, TId>;
     getListProps(): DataSourceListProps;
     getVisibleRows(): DataRowProps<TItem, TId>[];
     getSelectedRows(range?: VirtualListRange): DataRowProps<TItem, TId>[];
     getSelectedRowsCount(): number;
     reload(): void;
-    destroy(): void;
+    /**
+     * Activates IDataSourceView.
+     * After view activation, it becomes able to listen to updates.
+     */
+    activate(): void;
+    /**
+     * Deactivates IDataSourceView.
+     * After view deactivation, it becomes impossible to listen to updates.
+     */
+    deactivate(): void;
     loadData(): void;
+    clearAllChecked(): void;
     _forceUpdate(): void;
     selectAll?: ICheckable;
 };