+++ uui/components/buttons/Button.tsx
@@ -7,20 +7,30 @@ import { CountIndicator } from '../widgets';
 import css from './Button.module.scss';
 
 const defaultSize = '36';
-export type ButtonColor = 'accent' | 'primary' | 'critical' | 'secondary' | 'neutral';
-export type ButtonFill = 'solid' | 'outline' | 'ghost' | 'none';
 
-export interface ButtonMods {
+export type ButtonMods = {
+    /**
+     * Defines component size.
+     * @default '36'
+     */
     size?: ControlSize | '18';
-    fill?: ButtonFill;
-    color?: ButtonColor;
-}
-
-export type ButtonCoreProps = uuiButtonProps & {
-    size?: ControlSize | '18';
+    /**
+     * Defines component fill style.
+     * @default 'solid'
+     */
+    fill?: 'solid' | 'outline' | 'ghost' | 'none';
+    /**
+     * Defines component color.
+     * @default 'primary'
+     */
+    color?: 'accent' | 'primary' | 'critical' | 'secondary' | 'neutral';
 };
 
-export type ButtonProps = ButtonMods & Omit<uuiButtonProps, 'count' | 'indicator'>;
+/** Represents the 'Core properties' for the Button component, omitting the 'count' property. */
+export type ButtonCoreProps = Omit<uuiButtonProps, 'count'> & {};
+
+/** Represents the props for a Button component. */
+export type ButtonProps = ButtonMods & ButtonCoreProps;
 
 export function applyButtonMods(mods: ButtonProps) {
     return [